% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSdata.R
\docType{data}
\name{AEdemand}
\alias{AEdemand}
\title{Accident and Emergency demand in the UK}
\format{An object of class \code{ts}.}
\usage{
AEdemand
}
\description{
Weekly demand of Accident & Emergency departments in the UK, 
from 7 November 2010 to 7 June 2015.
}
\examples{
library(ggplot2)
autoplot(AEdemand, xlab="Year", ylab="Demand ('000)") +
  ggtitle("Accident & Emergency Demand in the UK")
}
\references{

}
\keyword{datasets}

