% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thief.R
\name{thief}
\alias{thief}
\title{Temporal hierarchy forecasting}
\usage{
thief(y, m = frequency(y), h = m * 2, comb = c("struc", "mse", "ols",
  "bu", "shr", "sam"), model = c("ets", "arima", "theta", "naive", "snaive"),
  ...)
}
\arguments{
\item{y}{Time series input}

\item{m}{Seasonal period}

\item{h}{Forecast horizon}

\item{comb}{Combination method of temporal hierarchies, taking one of the following values:
\describe{
  \item{"struc"}{Structural scaling - weights from temporal hierarchy}
  \item{"mse"}{Variance scaling - weights from in-sample MSE}
  \item{"ols"}{Unscaled OLS combination weights}
  \item{"bu"}{Bottom-up combination -- i.e., all aggregate forecasts are ignored.}
  \item{"shr"}{GLS using a shrinkage (to block diagonal) estimate of residuals}
  \item{"sam"}{GLS using sample covariance matrix of residuals}
}}

\item{model}{Model used for forecasting each aggregation level:
\describe{
  \item{"ets"}{exponential smoothing, using the \code{\link[forecast]{ets}} function.}
  \item{"arima"}{arima, using the \code{\link[forecast]{auto.arima}} function.}
  \item{"theta"}{theta method, using the \code{\link[TStools]{theta}} function.}
  \item{"naive"}{random walk forecasts}
  \item{"snaive"}{seasonal naive forecasts, based on the last year of observed data.}
}}

\item{...}{Arguments to be passed to the time series modelling function 
(such as \code{ets} or \code{auto.arima}).}
}
\value{
forecast object for bottom level series.
}
\description{
Takes a time series as input and produces forecasts using
the temporal hierarchical approach of Athanasopoulos et al (2016).
}
\examples{
  z <- thief(MAPA::admissions)
  plot(z)

}
\author{
Rob J Hyndman and Nikolaos Kourentzes
}

