% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{thiefcombine}
\alias{thiefcombine}
\title{Combine temporal hierarchical forecasts}
\usage{
thiefcombine(forecasts, comb = c("struc", "mse", "ols", "bu", "shr", "sam"),
  mse = NULL, residuals = NULL)
}
\arguments{
\item{forecasts}{List of forecasts. Each element must be a time series of forecasts. 
The number of forecasts should be equal to k times the seasonal period for each series, 
where k is the same across all series.}

\item{comb}{Combination method of temporal hierarchies, taking one of the following values:
\describe{
  \item{"struc"}{Structural scaling - weights from temporal hierarchy}
  \item{"mse"}{Variance scaling - weights from in-sample MSE}
  \item{"ols"}{Unscaled OLS combination weights}
  \item{"bu"}{Bottom-up combination -- i.e., all aggregate forecasts are ignored.}
  \item{"shr"}{GLS using a shrinkage (to block diagonal) estimate of residuals}
  \item{"sam"}{GLS using sample covariance matrix of residuals}
}}

\item{mse}{A vector of one-step MSE values corresponding to each of the forecast series.}

\item{residuals}{List of residuals corresponding to each of the forecast models. 
Each element must be a time series of residuals}
}
\value{
time series of reconciled bottom level series.
}
\description{
Takes forecasts of time series at all levels of temporal aggregation 
and combines them using the temporal hierarchical approach of Athanasopoulos et al (2016).
}
\examples{
aggts <- tsaggregates(USAccDeaths)
fc <- list()
for(i in 1:length(aggts))
  fc[[i]] <- forecast(aggts[[i]], h=2*frequency(aggts[[i]]))$mean
z <- thiefcombine(fc)
library(ggplot2)
autoplot(tsaggregates(z))

}
\author{
Rob J Hyndman
}

